Separate classes to own files + Remove snapkit and set constraints
