+++ Client/Frontend/Widgets/TwoLineImageOverlayCell.swift
@@ -1,0 +1,161 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0
+
+import UIKit
+
+// TODO: Add support for accessibility for when text size changes
+class TwoLineImageOverlayCell: UITableViewCell, ReusableCell {
+
+    struct UX {
+        static let ImageSize: CGFloat = 29
+        static let BorderViewMargin: CGFloat = 16
+    }
+
+    var notificationCenter: NotificationProtocol = NotificationCenter.default
+
+    /// Cell reuse causes the chevron to appear where it shouldn't. So, we use a different reuseIdentifier to prevent that.
+    static let accessoryUsageReuseIdentifier = "temporary-reuse-identifier"
+
+    // Tableview cell items
+    private lazy var selectedView: UIView = .build { view in
+        view.backgroundColor = UIColor.theme.tableView.selectedBackground
+    }
+
+    lazy var containerView: UIView = .build { view in
+        view.backgroundColor = .clear
+    }
+
+    lazy var leftImageView: UIImageView = .build { imageView in
+        imageView.contentMode = .scaleAspectFit
+        imageView.layer.cornerRadius = 5.0
+        imageView.clipsToBounds = true
+    }
+
+    lazy var leftOverlayImageView: UIImageView = .build { imageView in
+        imageView.contentMode = .scaleAspectFit
+        imageView.clipsToBounds = true
+    }
+
+    lazy var stackView: UIStackView = .build { stackView in
+        stackView.spacing = 8
+        stackView.axis = .vertical
+        stackView.distribution = .fillProportionally
+    }
+
+    lazy var titleLabel: UILabel = .build { label in
+        label.textColor = .black
+        label.font = UIFont.systemFont(ofSize: 16, weight: .regular)
+        label.textAlignment = .natural
+        label.numberOfLines = 1
+    }
+
+    lazy var descriptionLabel: UILabel = .build { label in
+        label.textColor = UIColor.Photon.Grey40
+        label.font = UIFont.systemFont(ofSize: 14, weight: .regular)
+        label.textAlignment = .natural
+        label.numberOfLines = 1
+    }
+
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        setupLayout()
+        setupNotifications(forObserver: self,
+                           observing: [.DisplayThemeChanged])
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private func setupLayout() {
+        separatorInset = UIEdgeInsets(top: 0,
+                                      left: UX.ImageSize + 2 * UX.BorderViewMargin,
+                                      bottom: 0,
+                                      right: 0)
+        selectionStyle = .default
+        stackView.addArrangedSubview(titleLabel)
+        stackView.addArrangedSubview(descriptionLabel)
+
+        containerView.addSubview(leftImageView)
+        containerView.addSubview(stackView)
+        containerView.addSubview(leftOverlayImageView)
+
+        contentView.addSubview(containerView)
+        bringSubviewToFront(containerView)
+
+        NSLayoutConstraint.activate([
+            containerView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            containerView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            containerView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
+            containerView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
+                                                    constant: -16),
+
+            leftImageView.leadingAnchor.constraint(equalTo: containerView.leadingAnchor,
+                                                   constant: 16),
+            leftImageView.centerYAnchor.constraint(equalTo: containerView.centerYAnchor),
+            leftImageView.heightAnchor.constraint(equalToConstant: 28),
+            leftImageView.widthAnchor.constraint(equalToConstant: 28),
+            leftImageView.trailingAnchor.constraint(equalTo: stackView.leadingAnchor,
+                                                   constant: -16),
+
+            leftOverlayImageView.trailingAnchor.constraint(equalTo: leftImageView.trailingAnchor,
+                                                          constant: 8),
+            leftOverlayImageView.bottomAnchor.constraint(equalTo: leftImageView.bottomAnchor,
+                                                        constant: 8),
+            leftOverlayImageView.heightAnchor.constraint(equalToConstant: 22),
+            leftOverlayImageView.widthAnchor.constraint(equalToConstant: 22),
+
+            stackView.topAnchor.constraint(equalTo: containerView.topAnchor,
+                                           constant: 8),
+            stackView.bottomAnchor.constraint(equalTo: containerView.bottomAnchor,
+                                              constant: -8),
+            stackView.trailingAnchor.constraint(equalTo: containerView.trailingAnchor,
+                                                    constant: -8),
+        ])
+
+        selectedBackgroundView = selectedView
+        applyTheme()
+    }
+
+    func configure() {
+
+    }
+
+    func applyTheme() {
+        let theme = BuiltinThemeName(rawValue: LegacyThemeManager.instance.current.name) ?? .normal
+        if theme == .dark {
+            self.backgroundColor = UIColor.Photon.Grey80
+            self.titleLabel.textColor = .white
+            self.descriptionLabel.textColor = UIColor.Photon.Grey40
+            self.selectedView.backgroundColor = UIColor.theme.tableView.selectedBackground
+        } else {
+            self.backgroundColor = .white
+            self.titleLabel.textColor = .black
+            self.descriptionLabel.textColor = UIColor.Photon.DarkGrey05
+            self.selectedView.backgroundColor = UIColor.theme.tableView.selectedBackground
+        }
+    }
+
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        self.selectionStyle = .default
+        separatorInset = UIEdgeInsets(top: 0,
+                                      left: UX.ImageSize + 2 * UX.BorderViewMargin,
+                                      bottom: 0,
+                                      right: 0)
+        leftImageView.image = nil
+        applyTheme()
+    }
+}
+
+extension TwoLineImageOverlayCell: Notifiable {
+    func handleNotifications(_ notification: Notification) {
+        switch notification.name {
+        case .DisplayThemeChanged:
+            applyTheme()
+        default:
+            break
+        }
+    }
+}
